Use walk method to have better control ove copied files
Previously the config path from the blueprint setting would be
processed, but after this change only those parts of the filepath which
are contained within the blueprint config directory are processed.
This enables better separation of concerns which enabled various
unwanted behaviour to be discarded.

These are still separation of concern issues to be dealt with.